rename register routes for consistency
